fix BTHome for irregular devices (#89)
Looks good.

Just for my future reference. The original 0x40 direct comparison was because that was the only option for BTHome when this driver was developed. Now, that byte has many more options. This change filters the byte to just check the relevant options for the needs of this driver.